Create help-meif any
I just want to be free from this prison that feels like my mind.
I want to be seen by the world 
I crave @abeltesfaye's attention more than most things though.
I long so strongly to get to know all of him in person.
I only want to be put into the life that he sees, breathes, and leaves. 
I long to have his head in my arms.
To give him a love that he deserves, and to find out if that love will be returned.